+++ addons/godot_state_charts/compound_state.gd
@@ -231,8 +231,11 @@ func _get_configuration_warnings() -> PackedStringArray:
 		if child is StateChartState:
 			child_count += 1
 
-	if child_count < 2:
-		warnings.append("Compound states should have at two child states.")
+	if child_count < 1:
+		warnings.append("Compound states should have at least one child state.")
+	
+	elif child_count < 2:
+		warnings.append("Compound states with only one child state are not very useful. Consider adding more child states or removing this compound state.")
 		
 	var the_initial_state = get_node_or_null(initial_state)